Move generated binding position check to src/utils
